HW5

+++ src/services/contacts.js
@@ -7,9 +7,11 @@ export const getContacts = async ({
   perPage = 10,
   sortBy = '_id',
   sortOrder = sortList[0],
+  userId,
 }) => {
   const skip = (page - 1) * perPage;
-  const data = await ContactsCollection.find()
+
+  const data = await ContactsCollection.find({ userId })
     .skip(skip)
     .limit(perPage)
     .sort({ [sortBy]: sortOrder });
@@ -30,19 +32,19 @@ export const getContacts = async ({
   };
 };
 
-export const getContactById = (contactId) =>
-  ContactsCollection.findOne({ _id: contactId });
+export const getContactById = (contactId, userId) =>
+  ContactsCollection.findOne({ _id: contactId, userId });
 
 export const addContact = (payload) => ContactsCollection.create(payload);
 
-export const updateContact = async (contactId, payload) => {
+export const updateContact = async (contactId, payload, userId) => {
   const updatedContact = await ContactsCollection.findOneAndUpdate(
-    { _id: contactId },
+    { _id: contactId, userId },
     { $set: payload },
   );
 
   return updatedContact;
 };
 
-export const deleteContactById = (contactId) =>
-  ContactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContactById = (contactId, userId) =>
+  ContactsCollection.findOneAndDelete({ _id: contactId, userId });
